Add files via upload
So I tinkered about with the multiplayer options n stuff. tried to make it more clear whos go it was and when it is a players go only their ships are visible. lemme know if I need to fix anything but hopefully you can incorporate it into the main code :)