Add files via upload
this file is ready for next weeks meeting. It has the multiplayer options, and AI difficulty options. easy and medium are pretty much random (medium has a slightly probability) and hard will be random till it hits a boat in which case it will lock on till the ship is destroyed then be random until it locks onto another.